requirements.txt
transformers>=4.40.0
datasets>=2.18.0
evaluate>=0.4.0
sacrebleu>=2.4.0
sentencepiece>=0.2.0
accelerate>=0.28.0
pytorch-lightning>=2.2.0
torch>=2.1.0
openpyxl>=3.1.0
pandas>=2.0.0
pyyaml>=6.0.1

peft>=0.10.0
bitsandbytes>=0.43.0
